requirements.txt
asn1crypto==0.22.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
clearbit==0.1.5
cryptography==1.9
Django==1.11.2
django-vote==2.1.4
djangorestframework==3.6.3
idna==2.5
pbkdf2==1.3
pycparser==2.17
pyhunter==0.2.2
PyJWT==1.5.2
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
